# docs/swagger.yaml
definitions:
...
  models.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon_url:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.CompanyInfoResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      facebook_url:
        type: string
      id:
        type: string
      instagram_url:
        type: string
      linkedin_url:
        type: string
      logo_url:
        type: string
      name:
        type: string
      phone:
        type: string
      twitter_url:
        type: string
      updated_at:
        type: string
      website_url:
        type: string
      youtube_url:
        type: string
    type: object
  models.CreateCategoryRequest:
    properties:
      description:
        maxLength: 500
        type: string
      icon_url:
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      sort_order:
        type: integer
    required:
    - name
    type: object
...
  models.Discount:
    properties:
      amount:
        description: Percentage or fixed amount
        type: number
      created_at:
        type: string
      description:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_quantity:
        description: Maximum quantity for discount (0 = unlimited)
        type: integer
      min_quantity:
        description: Minimum quantity for discount
        type: integer
      name:
        type: string
      tier:
        $ref: '#/definitions/models.EventTier'
      tier_id:
        description: If discount applies to specific tier
        type: string
      type:
        description: percentage, fixed_amount, buy_x_get_y
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
      usage_limit:
        description: 0 = unlimited
        type: integer
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
  models.Event:
    properties:
      address:
        type: string
      admin_remark:
        type: string
      available:
        type: integer
      banner_image:
        type: string
      cancel_reason:
        type: string
      cancelled_at:
        type: string
      capacity:
        minimum: 1
        type: integer
      category:
        description: Array of category tags
        items:
          type: string
        type: array
      commission_rate:
        description: Platform commission percentage (0-100)
        type: number
      created_at:
        type: string
      description:
        description: HTML content
        type: string
      discounts:
        items:
          $ref: '#/definitions/models.Discount'
        type: array
      end_date:
        type: string
      id:
        type: string
      is_cancelled:
        type: boolean
      is_featured:
        description: Featured event flag
        type: boolean
      location:
        description: Keep for backward compatibility
        type: string
      organizer:
        $ref: '#/definitions/models.User'
      organizer_id:
        type: string
      price:
        description: Base price for backward compatibility
        minimum: 0
        type: number
      promocodes:
        items:
          $ref: '#/definitions/models.Promocode'
        type: array
      sales_status:
        description: active, paused, stopped
        type: string
      start_date:
        type: string
      status:
        description: draft, pending, approved, held, rejected, cancelled
        type: string
      tiers:
        description: Relations
        items:
          $ref: '#/definitions/models.EventTier'
        type: array
      timezone:
        type: string
      title:
        type: string
      updated_at:
        type: string
      venue_name:
        type: string
    required:
    - capacity
    - end_date
    - price
    - start_date
    - title
    type: object
...
  models.EventTier:
    properties:
      available:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      gst:
        description: GST percentage
        type: number
      id:
        type: string
      is_active:
        type: boolean
      price:
        type: number
      quantity:
        type: integer
      sales_end:
        type: string
      sales_start:
        type: string
      sold:
        type: integer
      sort_order:
        description: For ordering tiers
        type: integer
      tier_name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Organization:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      logo_url:
        type: string
      members:
        items:
          $ref: '#/definitions/models.User'
        type: array
      name:
        type: string
      organizer:
        $ref: '#/definitions/models.User'
      organizer_id:
        type: string
      updated_at:
        type: string
      website_url:
        type: string
    type: object
...
  models.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      resource:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
    type: object
...
  models.Promocode:
    properties:
      amount:
        description: Percentage or fixed amount
        type: number
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_discount:
        description: Maximum discount amount (0 = unlimited)
        type: number
      min_amount:
        description: Minimum order amount
        type: number
      name:
        type: string
      tier:
        $ref: '#/definitions/models.EventTier'
      tier_id:
        description: If promocode applies to specific tier
        type: string
      type:
        description: percentage, fixed_amount
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
      usage_limit:
        description: 0 = unlimited
        type: integer
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.UpdateCategoryRequest:
    properties:
      description:
        maxLength: 500
        type: string
      icon_url:
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      sort_order:
        type: integer
    type: object
  models.UpdateCompanyInfoRequest:
    properties:
      address:
        maxLength: 500
        type: string
      description:
        maxLength: 2000
        type: string
      email:
        type: string
      facebook_url:
        type: string
      instagram_url:
        type: string
      linkedin_url:
        type: string
      logo_url:
        type: string
      name:
        maxLength: 200
        minLength: 2
        type: string
      phone:
        maxLength: 20
        minLength: 7
        type: string
      twitter_url:
        type: string
      website_url:
        type: string
      youtube_url:
        type: string
    required:
    - email
    - name
    type: object
...
  models.User:
    properties:
      account_status:
        description: active, inactive, suspended
        type: string
      admin_remark:
        type: string
      country_code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_email_verified:
        type: boolean
      last_name:
        type: string
      organization:
        $ref: '#/definitions/models.Organization'
      organization_id:
        type: string
      organizer_status:
        description: inactive, pending, approved, rejected
        type: string
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
    type: object
...
  utils.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      fields:
        description: For validation errors
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/utils.ErrorInfo'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A scalable REST API for event ticketing system with secure authentication
    and role-based access control (RBAC)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Timro Ticket API
  version: "1.0"
paths:
  /api/v1/admin/categories:
    get:
      consumes:
      - application/json
      description: Get all categories including inactive ones for admin management
      produces:
      - application/json
      responses:
        "200":
          description: Categories list
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CategoryResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all categories (Admin)
      tags:
      - Admin Management
    post:
      consumes:
      - application/json
      description: Create a new event category
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Category created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CategoryResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Category already exists
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Create new category (Admin)
      tags:
      - Admin Management
  /api/v1/admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete category (Admin)
      tags:
      - Admin Management
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CategoryResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Category name already exists
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Update category (Admin)
      tags:
      - Admin Management
  /api/v1/admin/company-info:
    get:
      consumes:
      - application/json
      description: Get company information for admin management
      produces:
      - application/json
      responses:
        "200":
          description: Company information
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CompanyInfoResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get company information (Admin)
      tags:
      - Admin Management
    put:
      consumes:
      - application/json
      description: Update website company information
      parameters:
      - description: Company information data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCompanyInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Company information updated
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CompanyInfoResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Update company information (Admin)
      tags:
      - Admin Management
...
  /api/v1/admin/events/{id}/featured:
    put:
      consumes:
      - application/json
      description: Mark/unmark an event as featured
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Featured status
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Event featured status updated
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Event'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Event not found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Toggle event featured status (Admin)
      tags:
      - Admin
...
securityDefinitions:
  ApiKeyAuth:
    description: 'JWT token authentication. Use the ''Bearer'' prefix followed by
      a space and the access token. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"